bsp-test-helper: fixed mount to usb pendrive
We cannot decide which device is usb pendrive of usb hdd
since they are storage devices detected by kernel. In case
we detect the device have same size, we cannot decide which
device should be mount. So just let usb hdd to /dev/sda1 and
usb pendrive to /dev/sdb1.

Signed-off-by: Mohamad Noor Alim Hussin <mohamad.noor.alim.hussin@intel.com>